Completed shoot() !
The script shoot() now works! It takes three arguments in this order:
rate_of_fire, bullet_speed, bullet_mass.

Notably, if rate_of_fire is a LOW number, the player will shoot FASTER,
and vice versa.

I left a huge number of comments as I worked. You'll definitely notice
the changes in the code.

Currently, obj_testplayer calls shoot() with magic numbers 20, 10, 1.
Since 10 and 1 are the speed and mass of obj_bullet, respectively, the
behavior of the code remains the same.