Grapple!
Buggy but close version of grappling hooks.  Grapple is shot with space
and deactivated with space.  There are 2 bugs.  1 the grapple is not
"active" in part of the region of the room based on where it was shot
from.  2 when you reenter a part of the room where grapple works, the
rope length is redefined.  Otherwise works awesome!  Also need to work
out some interactions with hitting walls